noticeme: a quick static check for unused values
Initial public release.

This tool is in the same genre as errcheck and unusedresult, and
provides a quick sanity check for "are values of this type being
used". It was written to check for lingering/forgotten cases
from a time when some operations were in-place mutator methods
rather than methods that would sometimes yield new objects
as results.